+++ src/components/SavedNewsPage/SavedNewsPage.js
@@ -7,27 +7,9 @@ import Navigation from '../Navigation/Navigation';
 import ButtonWired from '../ButtonWired/ButtonWired';
 import SavedNewsHeader from '../SavedNewsHeader/SavedNewsHeader';
 import NewsCardList from '../NewsCardList/NewsCardList';
+import { getKeywords } from '../../utils/keywordUtils';
 import logoutImageDark from '../../images/logout-dark.svg';
 import './SavedNewsPage.css';
-
-const formatKeyword = (keyword) => keyword.slice(0, 1).toUpperCase() + keyword.slice(1).toLowerCase();
-
-const getKeywords = (articles) => {
-  const counts = {};
-
-  articles.forEach(({ keyword }) => {
-    const formattedKeyword = formatKeyword(keyword);
-
-    if (counts[formattedKeyword]) {
-      counts[formattedKeyword] += 1;
-      return;
-    }
-
-    counts[formattedKeyword] = 0;
-  });
-
-  return Object.keys(counts);
-};
 
 const SavedNewsPage = ({ articles, onArticleUnsave, onLogout }) => {
   const currentUser = React.useContext(CurrentUserContext);
