+++ src/utils/keywordUtils.js
@@ -1,0 +1,23 @@
+export const formatKeyword = (keyword) => (
+  keyword
+    .split(' ')
+    .map((word) => word.slice(0, 1).toUpperCase() + word.slice(1).toLowerCase())
+    .join(' ')
+);
+
+export const getKeywords = (articles) => {
+  const counts = {};
+
+  articles.forEach(({ keyword }) => {
+    const formattedKeyword = formatKeyword(keyword);
+
+    if (counts[formattedKeyword]) {
+      counts[formattedKeyword] += 1;
+      return;
+    }
+
+    counts[formattedKeyword] = 0;
+  });
+
+  return Object.keys(counts);
+};
